move service-observer to services/core
